Fix add spent time by introducing submit callback
Commit 96d5c823544256f19e9cc7c429034883590d2670 introduced a bug where adding spent time resulted in an exception caused by the application flow.

The change itself is an improvement, however, the architecture within "add_spent_time_controller" did not support this improvement because it relied on StringVars dynamically mutating.

Since the function "add_spent_time" started the main loop for tkinters UI, the widgets were in an uninitialized state at the time of submission, ultimately leading to the exception being thrown.